split off sequential (not working) to separate file
